refactor : 패키지 변경
- 컨트롤러
- 도메인
- 뷰
- 유틸 (도메인과 뷰에서 둘다 사용)
  : validation패키지, GameResult,BridgeRule
