+++ src/main/java/bridge/util/BridgeRule.java
@@ -1,4 +1,4 @@
-package bridge;
+package bridge.util;
 
 public class BridgeRule {
