+++ src/main/java/bridge/util/validation/ViewValidator.java
@@ -1,12 +1,10 @@
-package bridge.validator;
-
-import static bridge.validator.Exception.ERROR_MESSAGE_FOR_NOT_NUMBER;
+package bridge.util.validation;
 
 public class ViewValidator {
     public static void isNumber(String input) {
         if (input.chars().allMatch(Character::isDigit)) {
             return;
         }
-        throw new IllegalArgumentException(ERROR_MESSAGE_FOR_NOT_NUMBER);
+        throw new IllegalArgumentException(Exception.ERROR_MESSAGE_FOR_NOT_NUMBER);
     }
 }
