+++ src/main/java/bridge/view/InputView.java
@@ -1,6 +1,6 @@
 package bridge.view;
 
-import bridge.validator.ViewValidator;
+import bridge.util.validation.ViewValidator;
 import camp.nextstep.edu.missionutils.Console;
 
 /**
