+++ src/main/java/bridge/view/OutputView.java
@@ -1,7 +1,7 @@
 package bridge.view;
 
 import bridge.domain.BridgeGame;
-import bridge.domain.GameResult;
+import bridge.util.GameResult;
 
 /**
  * 사용자에게 게임 진행 상황과 결과를 출력하는 역할을 한다.
